rmd-22827 Move detailed report logic to DO
